Change the main build/development branch to make corrections because of missing steps in the release documentation. Ensure that hte mven plugin references the current snapshot version of the tool and not the prior version, ensure that the tool uses the newly released version of ANTLR to build the next development version, remove redundant dependencies from the master pom.xml, and update the BUILD.txt file so we do not make this mistake next time.
Because the development version of ANTLR now uses release 3.1.3
of the plugin, it needs to relocate the v3 grammars under the
directory src/main/antlr3 and not src/main/antlr as the v2 plugin
uses src/main/antlr by default.

Change the newly released version of the maven plugin to use
the release 3.1.3 version of the tool and not the 3.1.3-SNAPSHOT
version. This causes the maven plugin to be release 3.1.3-1 :-(

[git-p4: depot-paths = "//depot/code/antlrworks/main/": change = 5943]